% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataIQ}
\alias{dataIQ}
\title{dataIQ: Data file for examples used in EPSY 905}
\format{
A data frame with 5 observations and 2 variables
\describe{
  \item{perf}{Performance}
  \item{iq}{IQ}
}
}
\usage{
dataIQ
}
\description{
A dataset containing the perf (Performance), iq (IQ) of 5 hypothetical people
}
\keyword{datasets}
