% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{homework}
\alias{homework}
\title{A Function to check homework of ESRM 64504}
\usage{
homework()
}
\description{
This function allows you to check out the progress of homework
}
\examples{
homework()
}
\keyword{homework}
